MH-13208, Create a short contributor guide
This patrch tries to create a short contributor guide, linking the
development process documentation but avoiding its lengty explanations.
The guide is in Github's default contributing guideline formnat and
should hence be pointed out by Github:

  https://help.github.com/articles/setting-guidelines-for-repository-contributors/

The guide intentionally tries to avoid mentioning any special cases and
points this out to readers. If necessary, they can still be found in the
developer documentation but most of the time, just sticking to this
guide should be fine.

The guide is meant especially for new, maybe even one-time contributors
which likely would not have read the long documentation before. Using
this, they should be able to quickly get an overview of how to create
proper pull requests for Opencast.