Remove non-functional automerge
Unfortunately, dependabot refuses to adhere to another bot's command to
automatically merge a pull request, resulting in bot's talking to each
other on dependabot's ESLint updates but no automatic merges.

This patch removes the non-functional automerge workflow. It doesn't
make sense to keep it.

We can explore some alternatives instead. There seem to be some GitHub
functionalities to do something like this:

- https://github.com/marketplace/actions/merge-me

I will open an issue as a reminder.