Limit Deployment to Opencast Community Repository
The deployment of documentation and translation keys as well as the
update of translations would run on all repositories, failing on all but
the community repository due to missing repository secrets. This will
cause unnecessary failure notices to be sent to the owners.

This patch changes the behavior, running the workflows only within the
Opencast GitHub organization