No Matrix Build on Opencast 8
On Opencast 9+, builds are run against both Java 8 and Java 11. Since we
dont want to build and upload the distribution twice, the upload steps
are limited to Java 11 only.

This does not work for Opencast 8 and this patch removes this additional
check.

This patch should be reverted when merged into `r/9.x`.