Added a new column "Notes" to Admin UI Event Table
A new column is added to the events list in the admin UI, which display text areas for users to make notes in. Text typed into these areas will be persistent and visible to other users.

"Notes" are saved as comments and thus leverage the comment backend. They therefore also appear in the comments view. All unresolved comments with the reason "Notes in Admin UI" will appear in the Notes column.

Only translated for en_US and de_DE.