Deactivate Parallel Builds in CI
This patch deactivates the recently activated parallel builds in the CI
system (GitHub Actions). Speed is not so much a concern there since we
need to wait at least to the next Technical Meeting for approval anyway.
Thus, we do not gain much by activating this.

But parallel Maven builds make identifying reasons for errors
incredibly hard. Especially if they happen in slow building modules
(e.g. in some JavaScript builds). Before you would just look at the end
of the build log. Now you need to scroll through thousands of lines of
logs to identify the root cause for a build failure. That is extremely
annoying as a reviewer.