Fix Date Time Picker
This patch fixes the interface of the date time picker which was missing
all its style attributes since they were not properly included in bower
and was additionally partially overwritten by random other attributes.

This also caused hidden fields like microseconds to show up which was
not helpful.

To make things even simpler, this also switches from the slider
interface for times to the selector interface, which is much more
common.