Automated Release Title
This patch should actually set the title in automatically generated
GitHub releases. The patch [is based on random information I found in
this thing called the internet](https://github.community/t/how-to-get-just-the-tag-name/16241/6).

I did not test this but it looks sensible and shouldn't make anything
worse. Let's just try this?