Auto-update JavaScript Test libraries
This patch adds an auto-update for the JavaScript dependencies of
engage-theodul-core which are eslint and other test libraries.

This should mean that we will usually no longer need to manually merge
the Dependabot pull requests for these modules.

This is done *only* for test libraries. No libraries which will actually
end up in Opencast itself are updated.

If the tests and thus the workflow fail, let's say eslint introduced new
default rules we do not comply with, this will eventually be picked up
by Dependabot which will then still create a pull request for us to
handle manually.

This is intentionally done for one module only to check if this works
properly. If it does, we can extend it to other modules which handly
only tests via npm.