Adapted Asset Upload
The options for the asset upload in event details and new event wizard, as well as source upload in new event wizard, are now fetched via API from the opencast configuration. The new event summary page has also been adapted to properly display the names of the asset and source types.