Remove red check mark when editing metadata
When editing meta data in the admin UI, a red check mark is shown on
changed (dirty/unwritten) fields. Red is not distinguishable from green
for some people, and the check mark as an indicator for unsaved changes
is not obvious.

Since the check marks do not mean anything anymore except that the user
finished their entry (which they should obviously know), this pull
request removes the check marks entirely.

If users try leaving the dialog with unsaved changes, a warning is
displayed anyway.

This fixes #3738