Neutral Modification Indicator
This patch adds a neutral modification indicator to the metadata editor in the admin interface. This uses the existing
mechanism to detect changes – it doesn't improve the algorithm, nor does it make it worse – but switches to a neutral
indicator for identifying changes from the read (Why read? Is it an error) check mark (Why a check mark? Do we want to
indicate success?) it was before.