Fixed Bulk Schedule Update
On the edit page of the udate scheduled event modal, the series dropdown and the weekday selection have been fixed.
On the summary page of the modal, the way series name and weekday are displayed have been fixed.
The request data for the API request for updating the event schedules has been adapted, so that the metadata and schedule are properly updated.