Remove Graphviz from Documentation
We already have a modern alternative to generating the somewhat blurry
graphs via the graphviz markdown extension and so this patch removes the
last remains of the Graphviz diagrams.

The reason for pushing that is that the old integration broke with the
release of new Python Markdown interpreter which causes our entire
documentation builds to be broken.

This patch replaces the graphviz diagrams with meremaid diagrams if
possible. The somewhat complex git branching models, I converted to
SVGs. Maybe we can use meremaid's GitGraph feature for this at some
point in the future.