Do not update default player components on legacy branches
This patch removes the configuration that tells Dependabot to update
components of Paella Player 6 – Opencast's default player – on legacy
branches.

The reasoning is that we know that dependency update pull requests are
often not well tested but get merged nonetheless. Doing so on `develop`
gives us time to react if something breaks. Doing this on a release
branch comes with a high risk of breaking the default player in a
release. Doing so in a legacy branch even increases the risk since less
people will test these.

There is usually no need for updating dependencies at all in the release
branches unless it is a security update or the dependency is the Player
itself, which should get proper attention and manual testing.

I don't mind auto-updating Paella Player 7, but only since that one is
not the default and specifically marked as “beta”.