Fix uploading videos using the proxy server
The proxy server was basically calling `toString` on the request body
which silently introduced replacement characters (�)
for every byte sequence which is not valid UTF-8,
which--it turns out--video files (and binary files in general)
contain a lot of!
This led to the uploaded video files reaching the proxied Opencast
to be corrupt, obviously.

This is no problem, thankfully,
since the `request{,-digest}` libraries can also deal with `Buffer`-s.
However, due to **how exactly** they do it, we have to take care
to only pass the body buffer if it's not empty. 🤷