Modernize redux: statisticsSlice
Switching to redux toolkit for getting statistics for the statistics view.

This breaks event and series statistics! Only organization
statistics should remain intact. Event and series statistics
will have to be fixed when their redux code is modernized.

Also this is largely untested because getting statistics
into Opencast is apparently a major pain.