Replace react-smooth-dnd with react-beautiful-dnd
While react-smooth-dnd works perfectly fine, it is
also unmaintained and will cause problems when
updating to new major versions of react. This
switches to a currently maintained library, which is
a lot more boilerplate, but also designed for our
use case (the list where you can edit which columns
are displayed in a table and in what order).

Also fixes a bug in the EditTable modal where changes
made to the sorting (without saving) would revert
themselves after a couple seconds.