Avoid crashing by not throwing error
Potentially controversial fix intended to keep nodemon from
crashing. Instead of having the proxy server
throw certain kinds of errors, this changes it to merely logging
them. The throwing in and of itself is likely not the root cause
of the nodemon crashing issue, but removing it seems to alleviate
the issue nonetheless.

I'm not certain on what we should ideally do with the kinds
of errors that were previously thrown in the first place though.