Add notes column back in
Add a column called  "notes" to  the event table. This was
already present in the old admin ui. The column provides
a small text area where a user can take short on an event.
The notes are saved as a comment (and can thus also be
viewed in the event details).

The notes column is disabled by default, and needs to be
manually enabled. Only if it is enabled will additional
comment information be fetched from the backend,
which puts a bit of additional load on the Opencast admin
if you have a lot of comments.