Reduce state persisted by redux-persist
One of the reasons redux-persist is causing us so many issues
is that we persist the whole store. This means we persist
a lot of information we really should not, either because it's
info that should get cleared on page reload or because we
fetch the info on page load anyway (this is not an offline app).

This introduces redux-persist whitelisting, so we only persist
what we explicitely declare to. From what I gather, we want
to persist table filters and column arrangement, so state
pertaining to that should be whitelisted.

This also fixes a faulty import statement, but it likely won't
have any noticable effect (since autoMergeLevel2 is not deep
enough of a level to affect our nested states).