Fix sameday filter for tables
For e.g. the event table, you can filter by start date,
by setting two dates to filter between.
This was not always working properly, because no
heed was paid to the time value for those filter dates.
This sets the time values for the dates, so filtering
from 14.02 to 14.02 should now show all videos from
14.02 instead of no videos at all.