Fix statistics in events details
Statistics were commented out for statistics in the
events details tab. This comments the functionality
back in and fixes it by making use of the new
functionality offeered by statisticsSlice.ts.

Same as for the "Modernize Statistics" PR, I have not
tested this with actual statistics, because setting that
up is very hard for some reason.