Harmonize date rendering between read-only and read/write fields
Metadata fields, that is. This expands on ecc570b2314431a1fc6ed6f0ad1cf49841e7e5a4
(itself a late addition to #277) and fixes a time drift between
the read-only "Created" field and the non-edit-mode variant of the
Start field (the values of which are synchronized) caused by the two
methods of formatting interpreting the source date in different
timezones.