Automatically deploy pull requests
This patch automatically deploys pull requests with static mock data to
https://test.admin-interface.opencast.org similar to how the test
deployment of the Opencast Editor works.

Unfortunately, we cannot set a path prefix for mock data which means
that all deployments share the lattest set of mock data. But since they
shouldn't change very often, that's probably okay for this to get
started.