Statistics tab enabled by default
By default the statistics view is broken and you need additional
configuration. Therefor, it should be disabled by default in the admin
interface.

Unlike the old interface, this patch does not read the actual data from
statistics providers, but uses a configuration option to determine if
the statistics feature should be an option in the UI.

This fixes #357