Fix admin interface crashing when rendering filters
This fixes the admin interface crashing if there is no preset label for
a given filter value. This will be the case, for example, if you filter
for usernames.

Note that this is only reproducible with the mock data right now since
there seems to be a bug with the procied data that prevents rendering so
you never reach this code 🙈

This fixes #526