"Restore" ESLint support
I removed our leftover config in #700 with a promise to bring ESLint back.
Well this is it, for now!

This introduces an ESLint config file in the new flat file format,
which will be required for ESLint `>= 9`. It is already supported in version 8, though,
which we are still stuck on because of the (unmaintained) config we are/were using.
I briefly tried to port this project over to `@opencast/eslint-config-ts-react`,
but the code said "no," basically. ;P

This has a few unfortunate consequences. See #713. I plan to address these in due time, though.

Note that this does **not** introduce automatic/live checking of the code
during development. I only added a call to the linter to the build command,
in turn making CI fail when there are warnings.
I created #712 to track the live thing if we want that.