Get rid of Vite build warnings
There are quite a few errors like the following in the output of `npm run build`:

    Error when using sourcemap for reporting an error: Can't resolve original location of error.

They all relate to the `@mui` package(-s). This seems to be related to https://github.com/vitejs/vite/issues/15012.
I don't fully understand it, but apparently Rollup, which is used by Vite internally somehow,
doesn't preserve enough information to generate source maps properly.

This patch adds a plugin to do exactly that. We should probably remove it in the long run,
if the upstream issue(-s) ever get fixed. I created #708 to track that.

/cc https://github.com/opencast/opencast-admin-interface/pull/685