Fix type error for missing organization property
If an organization property does not exist, the admin interface
completely breaks in some cases. This seems to happen on
admin-interface.opencast.org right now:

```
Uncaught TypeError: u.toLowerCase is not a function
    qr Footer.tsx:25
    React 8
    D scheduler.production.min.js:13
    $ scheduler.production.min.js:14
Footer.tsx:25:80
```

What's interesting is that this only seems to happen with the compiled
code. Starting the development server, calling `toLowerCase()` just
works.

In any case, this patch fixes the issue by providing the default value
first, so that `toLowerCase()` is always called on a string.