Add a very restricted form of `no-tabs` to the ESLint config
This is in preparation to adopting `@opencast/eslint-config-ts-react`,
which includes `no-tabs`.

We use tabs for indentation in some places, but also for other things.
The indentation tabs are easy to get rid of automatically, if a bit annoying.
It is the **other** instances that are not so easily picked out,
and especially distinguished from valid tab usages (think literals).

So I'm getting these out of the way first.