Hopefully fix a crash provoking bug
Something was causing a crash when using the breadcrumb links
to navigate to another page in the path editing menu while
being in "edit mode" (i.e. in the process of adding a new page).
This resulted in a "Maximum update depth exceeded" error, crashing
the interface. The error hints at circular rendering dependendies,
but unfortunately I couldn't reproduce this myself.

But by attempting to do that I noticed another bug that may or
may not be related: The list of children was missing an entry.
I removed the code doing that, since it didn't seem to be necessary
anyway - this also removed a state setting call that might have
caused too many rerenders.