Seperate user and non-user roles
Changes the access policy tab of various
details modals. User roles get displayed in a
seperate table from the other roles, with user
name and email. This should make it more
intuitive to "give someone access to a video/
series". Functionality should not change,
the resulting ACLs will look the same as before.

Caveat 1:  If roles are sanitized, we cannot
reliably derive the user from their user
role. Therefore, this feature will be disabled if
user role sanitization is enabled in Opencast.

Caveat 2: I did not find a performant  way to
query for user information for each and every
user.
Therefore, if there are upwards of multiple
thousands of users, the access policy tab will
take several seconds to load.