Fix tobira notification bug
Fixes an issue where (a) the `not_found` notification
would not show up in the tobira tab, leaving it
completely empty. In addition, the emitted error was never
reset, which (b) also broke the tab for events that
do get found.

(a) happened because we now fetch the event data directly
when opening an event modal, to check the overall tobira
status. Switching tabs however removes the notification,
which is why it needs to be dispatched again when the
tobira tab is opened.
(b) was probably just an oversight on my end.