Make use of acl template matching
Make use of a variable we can get from the backend
that has been ignored up till now. It gives us a template id
based on the roles in the current acl. We can make use of
this in the event and series details to display the
"correct" acl template instead of the usual "Select a template"
text.

The backend does some best effort matching, which
can be made more lenient with the
"match.managed.acl.role.prefixes"
configuration option.