Add interrupting mechanism to fetching in pages
When unmounting a page, dispatched actions
will continue to run in the background. This could cause the
wrong data being loaded into the table, because
the dispatched actions from one page would
complete later than the dispatched actions from another page.
This is most obvious when switching from
series to events and then quickly switching back to series.
This patch adds an interrupting mechanism that should
prevent the "loadIntoTable" action to be dispatched if the component
was unmounted.