Ensure correct sorting params are used for fetching
Fetching data for a page would sometimes use the sorting params from
the previous page, as the "resource" variable had not yet updated in
the tableSlice. This could lead to wrong requests which then had to be
corrected by subsequent requests. So instead of relying on the
"resource" variable in the tableSlice, the fetching functions now
hardcode their respective resource.