Switch to @opencast/eslint-config-ts-react
Thanks to @ LukasKalbertodt , our Opencast eslint config for
typescript-react is now compatible with eslint 9.
High time to not only update to eslint 9, but also to replace our old
eslint config with @opencast/eslint-config-ts-react.
The  old config was a default  config provided by CreateReactApp,
which is not maintained anymore and we have no other reason to keep
it around.

That said, this patch only aims to update dependencies and introduce
@opencast/eslint-config-ts-react. It does not aim to fulfill the
numerous rules the project is now breaking, which is why most of those
rules are disabled. They can be gradually enabeled by future patches.