Fix event/series modal not wide enough
The tabs in the event/series details modals were wrapping around
in case there were too many of them,
due to the of the container being too
small. This fixes that by applying the
correct class id for those modals again,
which was lost due to multiple PRs.

Arguably the fixed widths here are the true problem and we should look
into making our modal css more dynamic,
but that's a task for another PR.