FIx wrong publications on page switch
This patch fixes a bug that happened when switching pages
in the events table. Events that had no publication would be displayed
as having publications, if the event in its place on the previous page
had publications. For example, if the first event on page 1 had
publications, and then you would switch to page 2 where the first event
did not have publications, the event would be displayed as having
publications.

This patch fixes that by firstly resetting publication information
if the information changes. It also limits enriching to only happen if
the publications have changed.