Allow tracks in event asset upload (#698)
* Allow tracks in event asset upload

Old logic dictates that media of type track
cannot be in the asset upload tab of the event
details or new event wizard. Since subtitles
are now tracks, this logic finally breaks.

Instead of relying on checking for type
track to differentiate between options
for source upload and asset upload, this
commit instead uses the key in the list
provider that is there for this exact purpose.

* Fix import statement

* Add new optional keys for displaying asset upload

Adds support for `showForNewEvents` and
`showForExistingEvents`, two keys that
specifically specify whether an asset upload
option should show up for new events or
existing events. Previously, an asset upload
option would always show for new and
existing events, but now this can be configured
in detail.