Apply default actions to default roles for new events/series
Fixes #1354.

An admin can configure actions that are automatically
added to each role you add to an ACL in the ACL tab
for series and events. However, they were
not applied to default user role the ACL
is initialized with. This adds the actions to
default user role.

Fix custom actions being lost on template switch for default user

When switching between ACL templates, we were keeping
the default user as intended, but the default user
lost their default actions (if configured). This fixes that.

Fix  acl tab config settings not applying

The acl config defaults were not always fetched for the acl tab.
This should fix that.