Open workflow details first
When clicking on the Workflow tab in the event details,
the user is presented with a list of workflows to select
from. Most of the time, the user will selected the latest
workflow. So this patch selects it for them automatically.
The list of workflows can still be accessed through the tab
hierarch navigation.