Make event metadata load faster by fetching series options asynchronously
If you have many series in your Opencast, any modal that displays
event metadata can take several seconds to load. This patch aims at
pushing the loading time back into millisecond terrritory no matter
how many series there are.

The problem is that when fetching event metadata from the backend,
the backend also fetch all series options, which can take too long.
So we change the backend to not do that, and change the frontend to
fetch the series options separately.

ideally we could also do this for other metadata fields with options
(notably contributors and presenters), but the endpoint lacks
the necessary filter capabilities.