Show unsaved changes warning on tab switch
When changing event/series metadata/acl in the event/series details
and not explicitly saving them, the changes are lost when switching to
another tab. This patch causes the browser to show a warning dialog
when a user tries to switch to a different tab while there are unsaved
changes.