Default template for "Create Event/Series"
Includes #1360. #1360 should be reviewed and merged first.

This allows an admin to configure a default template for new
events/series. Upon opening the ACL tab of the respective create modal,
the template will already be selected and the roles respective to the
template will be set.

A default template will not overwrite the default user role of the
creator that is added automatically.