Disable tooltips on tables due to performance issues
This commit disables tooltips for all table cells. For example,
when moving the mouse over the red X in the events table,
there will no longer be a tooltip showing. This is done to
imporve performance.

It appears that the mui tooltip component is not performant when
used in large quantities. According to this comment on github it is not
supposed to be:
https://github.com/mui/material-ui/issues/27057#issuecomment-880236030

I locally tested performance by profiling the rendering of the rows of the
series table. The table had 1000 entries which necessitated 4000 tooltip
components. Rendering the table with tooltips took about 3,7 seconds.
Rendering the table without tooltips took about 1,2 seconds. I think
this is way too much time for tooltips.

The goal of this commit is to be a quick fix for quick gains.
A more proper long term solution would let us keep the tooltips, by for
example working around the issue somehow or replacing mui tooltips with
a different library.