Remove useEffect hook validation from create modals
This hook would call upon the formik validation when
the user switched to a different tab in the create modal.
This is pointless because A. validation will automatically have
been called by formik before then and B. this hook would only
trigger after the tab has changed,
which would be too late anyway.